+++ src/main/java/com/booklify/service/IPaymentService.java
@@ -2,6 +2,7 @@ package com.booklify.service;
 
 import com.booklify.domain.Payment;
 import com.booklify.domain.enums.PaymentStatus;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface IPaymentService extends IService<Payment, Long> {
@@ -9,5 +10,5 @@ public interface IPaymentService extends IService<Payment, Long> {
     List<Payment> getAll();
     List<Payment> findByUser(Long userId);
     List<Payment> findByStatus(PaymentStatus status);
-    boolean processRefund(Long paymentId, double amount);
+    boolean processRefund(Long paymentId, BigDecimal amount);
 }
